requirements.txt
Pillow==6.2.1
PyQt5==5.14.0
PyQt5-sip==12.7.0
requests==2.18.4
